Finalize initial endpoint setup
