+++ .gitignore
@@ -1,3 +1,14 @@
 node_modules
 dist
 .cache
+.DS_Store
+
+svc/.idea/
+
+env/lib/python3.6/site-packages/__pycache__/
+
+.idea/
+
+env/
+
+*.pyc
