+++ svc/src/__init__.py
@@ -1,0 +1,8 @@
+
+from flask import Flask
+from flask_debugtoolbar import DebugToolbarExtension
+app = Flask('project')
+app.config['SECRET_KEY'] = 'random'
+app.debug = True
+toolbar = DebugToolbarExtension(app)
+from src.http.endpoints import *
