+++ svc/src/http/endpoints/__init__.py
@@ -1,0 +1,1 @@
+from .instructor import *
