+++ svc/src/http/response/instructor_list_response.py
@@ -1,0 +1,6 @@
+from .instructor_response import from_data
+
+
+def from_query(data):
+    return [from_data(item) for item in data]
+
