+++ svc/src/http/response/instructor_response.py
@@ -1,0 +1,5 @@
+keys = ['id', 'first_name', 'last_name', 'email']
+
+
+def from_data(data):
+    return dict(zip(keys, data))
