+++ svc/src/main.py
@@ -1,4 +1,4 @@
-from src.project import app
+from src import app
 
 
 if __name__ == '__main__':
